Fix Windows build warnings

+++ dartvlc/equalizer.hpp
@@ -91,13 +91,13 @@ public:
 	}
 
 	int createEmpty() {
-		this->equalizers[this->equalizers.size()] = new Equalizer();
-		return this->equalizers.size();
+		this->equalizers[static_cast<int>(this->equalizers.size())] = new Equalizer();
+		return static_cast<int>(this->equalizers.size());
 	}
 
 	int createMode(EqualizerMode mode) {
-		this->equalizers[this->equalizers.size()] = new Equalizer(mode);
-		return this->equalizers.size();
+		this->equalizers[static_cast<int>(this->equalizers.size())] = new Equalizer(mode);
+		return static_cast<int>(this->equalizers.size());
 	}
 
 private:
